scripts:Add generic layer generator
Add generic_layer.py script that generates base framework for a new
layer that will intercept all Vulkan functions. The generic layer is
not useful by itself, but is intended to speed up creation of a custom
layer by providing an initial framework that can be copied and
customized.

The generic header can be generated which includes function prototypes
and the dispatch table.
The generic cpp file can also be generated which includes function
bodies that call down the chain.