WIP: Test architecture for VulkAid layer
Layer itself is doing nothing, just intercepts AllocateCommandBuffers
and creates a state wrapper for command buffer.

This is a snapshot of this architecture where the state wrappers are
at the same level as the layer_data wrapper. This allows a single look-
up into the device_layer_data_map and a ptr to device_layer_data can
then be passed along to all of the levels.

Right now, however, this exposes struct wrappers to all levels, which
is not ideal. If we can have a single map look-up while also hiding the
wrapper structs from the last level, that would be ideal. Haven't found
a way to make that actually work yet, though.