vkaid:Migrate cmd buffer output to MVC pattern
More architectural experimenting. Created view and controller classes
to enable display of state according to Model-View-Controller pattern.
Not sure of best place for View and Controller to live so initially
adding them as as unique_ptrs in the dispatch layer and initialize them
at DeviceCreate time.

This example is still not too exciting as the initial StateView class
just prints out cmd buffer commands to stdout.